Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I tried to compose something coherent but somehow submitted something else.  Tim you replied to my first incoherent draft.
> 
> I get it that this is new, not yet used.  Still, ahh, it seems to be worth a release note that mentions why this was added.  Maybe.

MAVEN_OPTIONS is NOT an environmental variable, it is a shell variable and should therefore be lower-case. When I start using this function I will change refs from upper-case to lower-case. When all refs have been changed, I will be able to delete gerrit-push-patch.sh, but that is not a priority, it will happen at some point.

Patch-set: 1
Label: Code-Review=0