Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review-1
> 
> (1 comment)

Sourcing this script does nothing, it only defines some functions which cannot be used until this file has been merged and released (tagged). Currently no scripts source this script, let alone attempt to call any of the functions in the script. I do have another review out that uses another function (lf-activate()) from this file.

The purpose of the lf-set-maven-options() function is replace common-variables.sh. The way we are using this script is really creative (not in a good way). I was expecting to defend the need for that change when you review the changes that replaces the use of 'common-variables.h' with a call to lf-set-maven-options().

This change is just an improvement to a library that no-one is using yet. I want to merge this change before anyone starts using it. This change makes the code more compatible with shellcheck.

Patch-set: 3
Label: Code-Review=0