Update patch set 1
Patch Set 1:

(1 comment)

Anil, I think something is still needed in this shell script. Even though I put a conditional to in the deploy click command to check if "nexus" is in the url and if it is, run nexus specific functions, and if not, run s3 functions, this script sets the url.

Patch-set: 1
Reviewer: Gerrit User 325 <325@f2a2391d-0403-4fa5-b183-4056999a8a09>
Label: Code-Review=0