Update patch set 41
Patch Set 41:

I undeandrs that the tox job assumes a tox.ini file will be present in the project root directory.  In the O-RAN-SC project portal/ric-dashboard, which is based on Java, Aric helped me with adding appropriate file.  But in O-RAN-SC project ric-plt-a1, which is based on Python, there already is a tox.ini file there!  Does the tox-run for rtdv2 supply the appropriate arguments to limit tox to building the doc only?

Patch-set: 41