Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Ok I pip-install-ed yamllint and ran it. Got warnings about missing spaces before comments and overly long lines, but nada about repeated keys.  Maybe I have an old version?  Please advise exactly what & how to run so I see what you see.

Interesting. I was using 1.10.0, from apt on Ubuntu 18. I installed 1.18.0 from pip, and the errors went away. Given that it's valid code, and more of an issue with the linter than the code, it makes sense that they would change that. Ok, in that case, I say just remove the comments.

Patch-set: 4