Update patch set 1
Patch Set 1:

This new function has been created to create a 'user' venv that can be safely shared.

Note that I am attempting to install a complete list of packages that are required by the build scripts. There is current no 'version pinning'. When all packages are installed at once pip is able to meet version dependencies. One of the problems with 'pip' is that even if there are version errors in the install it will still return a non-zero status. For this reason I  archive the log of the pip install so it can be easily reviewed.

The goal here is to quit having multiple scripts creating their one venv or having multiple scripts installing packages in the user venv. Because of the nature of the global-jjb, we may call lf-create-user-vent() multiple times. The function will just return if the venv has already be created.

None of the functions defined in the bash library are currently being referenced. I am working on a another change that is under development that will calling theses functions. They have to been used to thoroughly test all these changes. The changes to this library need to be released/tagged to become available.

Aric: I believe TOX will be able to use the user venv created by lf-create-user-vent() when we move TOX from pyenv -> virtualenv.

Patch-set: 1
Reviewer: Gerrit User 182 <182@f2a2391d-0403-4fa5-b183-4056999a8a09>
Label: Code-Review=0