Update patch set 2
Patch Set 2:

> (1 comment)

That requires the sigul infra bits in place. Also I've tried using the  -s option never works for me with gpg2 on F30. therefore I skip that and sign them manually. for instance the detached signatures for some of the releases can actually be found on the release repo and wont require to move them into the logs. 

https://files.pythonhosted.org/packages/d8/de/dd3842546398d82a241c5e31e9fe01de5df882234e8ad53c1c5d23f1bdbe/lftools-0.24.0-py2.py3-none-any.whl.asc

I've read that < 4% of the packages on pypi are being signed atm. that said we should definetly consider doing that.

Patch-set: 2