Update patch set 4
Patch Set 4:

I can cope with either niet or yq, but I note that they are not drop-in replacements for each other.  For example, niet by default sets exit code 1 if a key is not found; yq requires you to use the -e flag to get the same behavior.

Patch-set: 4