use tox-venv and tox-pip-version
Caveat: now that we know that its virtualenv that was breaking
everything, we should look into just updating that and
perhaps finding a more standard solution.

tox calls system virtualenv which breaks the coala linter.
as well as bringing with it a a very old version of pip from system.

This job is failing now bc the coala runs are actually
working, I manually edited tox-install to add the
tox-venv and tox-pip-version packages.

This is the change that broke us. (when we moved to python3)
https://tox.readthedocs.io/en/latest/changelog.html#v3-10-0-2019-05-13
v3.10.0
Virtual environments created now no longer upgrade pip/wheel/setuptools
to the latest version. Instead the start packages after virtualenv
creation now is whatever virtualenv has bundled in. This allows faster
virtualenv creation and builds that are easier to reproduce. #448

Fix uses two modules tox-venv so that we use python3 -m venv
rather than virtualenv and tox-pip-version so we can upgrade
the pip version sice the upgrade is no longer done with tox.

https://pypi.org/project/tox-venv/
tox-venv is a plugin that uses Python 3’s builtin venv module for
creating test environments, instead of creating them with the virtualenv
package. For Python versions that do not include venv (namely 3.2 and
earlier), this package does nothing and reverts to tox’s default
implementation.

https://pypi.org/project/tox-pip-version/

changes to tox.ini as well
set pip_version = 19.3.1
and test it with a new script.
added a script to check pip version
also removed two stanzas from tox.ini that were I belive
python2 workarounds, in any case I don't
see their function at this time.

Signed-off-by: Aric Gardner <agardner@linuxfoundation.org>
Change-Id: I1a0d05ef55966e4262482db92896648cfd487488