use tox-venv
https://pypi.org/project/tox-venv/
tox-venv is a plugin that uses Python 3’s builtin venv module for
creating test environments, instead of creating them with the virtualenv
package. For Python versions that do not include venv (namely 3.2 and
earlier), this package does nothing and reverts to tox’s default
implementation.

And:
https://pypi.org/project/tox-pip-version/

changes to tox.ini as well..
added a script to check pip version

Signed-off-by: Aric Gardner <agardner@linuxfoundation.org>
Change-Id: I1a0d05ef55966e4262482db92896648cfd487488