Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> Hi Tim, I dont think we should remove pyenv without communicating to projects.
> 
> also it is working now, and providing its expected benefit.
> 
> Also, I would like to test fixing this with a simplier
> pip install --upgrade virtualenv

I am not proposing removing pyenv altogether, only the snippet of code that references it in tox-run.sh. The code in tox-run.sh doesn't  change the pyenv, it just sets some variables for using pyenv for tox. It 'works' with it because none of the things it does is required anymore.

Patch-set: 11