Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> It occurs to me now, far too late of course, just how very useful it would be to have lf-venv-create accept a list of packages to install, basically a combo of lf-venv-create and lf-venv-add. Please consider it.

I did think of that, and can still do it, we needed to get this merged and in the hands of developers. Too early to think of perfect, 'usable' was what I was looking for.

Thinking about it now, I am betting the code change would be less than 1 hr (based on the login from lf-venv-add()), testing is the most time-consuming part.

Patch-set: 10