Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> Last suggestion (I promise).  Instead of every script starting:
> 
>     lf-venv-create python3
>     lf-venv-add python3 pkg2 pkg3 pkg4
>     lf-venv-activate python3
> 
> maybe just possibly you could provide features so script peeps like me write just this:
> 
>     lf-venv python3 pkg2 pkg3 pkg4
> 
> which will create (if needed), install (if needed), and add to path.

You don't have to promise anything. Agreed the lf-venv-add() will be folded into lf-venv-create). lf-venv-activate() is a very special animal, so it needs to stand alone. Hard to explain (goes back to an 'unwritten law' of Unix), but I would be glad to do so in phone/meeting setting.

Patch-set: 10