Update patch set 7
Patch Set 7:

> Patch Set 7: Code-Review+2
> 
> Looks good, we need to consider that jenkins jobs version is now set in 3 places and how to deal with that going forward, but for now this looks like a solid iteration.

It needs to be set in three places (for now).
1. shared venv
2. user venv (specified in lf-ci-jobs.yaml)
3. tox.ini

When the user venv goes away, we will be back to two. Because we don't control where/when lf-venv-create() is going to get called, I don't see a way to drive the version from yaml/jenkins-jobs. I would be open to changing the way it is set for shared venv, what I have works, but that is not to say there is not a better way.

Patch-set: 7