Update patch set 2
Patch Set 2:

One note I've been meaning to bring up and this patch reminded me of it. Regarding the regexes, it doesn't take into account if someone votes at the same time as leaving the trigger. For example:

 --- Example ---
 Patch Set 2: Code-Review+2
 
 recheck
 --- Example ---

(This example uses Code-Review, but the situation is the same if Verified or any other labels are used too)

I've seen occassionally this happens and folks get confused because their retrigger did not run. Our regex here used to be extremely simple in that it just checks for something that ends in `(recheck|reverify)$` which allowed voting at same time. Since we started tightening the regexes here it's caused this one use case to no longer work.

Maybe it'd be worth documenting somewhere this minor case.

Patch-set: 2