Update patch set 8
Patch Set 8:

aric, thank you for the merge.  But if you can PLEASE find a minute, go back to my previous comment and try the test I suggested.  Testing triggering behavior is a total spray-and-pray operation for me.  But you all, with admin privs on the ORAN jenkins, can twiddle the job, post a comment and observe in far less time than it takes me just to write this comment.

I'm concerned there's a deeper problem with the project/branch/path trigger pattern. The changed regex was only a SWAG, and your test will prove that out (or fail it), valuable information either way. 

Thanks in advance.

Patch-set: 8