Update patch set 11
Patch Set 11:

(2 comments)

I have moved away from sharing a single venv. The venv's are created as needed in /tmp.

The ability to add new packages to an existing venv turned out to be quite difficult because of the occasional need to support versioned packages. I would be willing to get into this decisions in more depth if anyone is interested. Sorry for way this whole thing dragged on for so long.... I am really feeling pretty good about it.

Patch-set: 11