Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> (2 comments)
> 
> I have moved away from sharing a single venv. The venv's are created as needed in /tmp.
> 
> The ability to add new packages to an existing venv turned out to be quite difficult because of the occasional need to support versioned packages. I would be willing to get into this decisions in more depth if anyone is interested. Sorry for way this whole thing dragged on for so long.... I am really feeling pretty good about it.

Tim, why not have a single "static" venv for all our tooling, and then have the functionality to create additional venvs as needed for those edge cases that have pinned versions, etc.?

The whole point of this is to speed up builds, no? Just having an "immutable" tools env alone would do that. Ideally it would be part of the builder image itself for speed's sake, but having it built live is okay too, if that isn't possible.

Patch-set: 11