Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> > Patch Set 11:
> > 
> > (2 comments)
> > 
> > I have moved away from sharing a single venv. The venv's are created as needed in /tmp.
> > 
> > The ability to add new packages to an existing venv turned out to be quite difficult because of the occasional need to support versioned packages. I would be willing to get into this decisions in more depth if anyone is interested. Sorry for way this whole thing dragged on for so long.... I am really feeling pretty good about it.
> 
> Tim, why not have a single "static" venv for all our tooling, and then have the functionality to create additional venvs as needed for those edge cases that have pinned versions, etc.?
> 
> The whole point of this is to speed up builds, no? Just having an "immutable" tools env alone would do that. Ideally it would be part of the builder image itself for speed's sake, but having it built live is okay too, if that isn't possible.

We have no way to create a script that 'always' run before anything else. We do have the jenkins-init script but it has a own special set of problems. I have found examples where scripts needed to do special 'things' with their venv that are just fine for them, but we would not want to do in a 'shared' venv. Most of the time each script only needs one or two packages: lftools, jenkins-job-builder and TOX mostly. One or two packages takes less than 10secs to install. The set of packages currently installed with python-tools-install take almost 2mins and it is still not a complete set of packages

Making the builds faster was not a big reason for starting down this path. Speed is something I always take into consideration. I felt we needed better control of the python venvs to make their use more predictable. The current process uses a 'user' venv that was 'magically' added to end of the PATH made access to python packages very difficult to predict. Turns out, python has a special set of rules that only apply to a user venv's. You really had no idea where the commands were coming from. With the new venv, when you run pip, venv, python or lftools, you know where they come based on a review of the code.

Patch-set: 11