Update patch set 1
Patch Set 1:

> The Acumos project uses release branch names athena, boreas and
 > clio.  All are published to ReadTheDocs.io.
 > 
 > We just recently stopped publishing the "master" branch over to
 > ReadTheDocs.io.

This only deals with event type 'ref-updated', in cases we have to run the job on new tags being published. I think every project has diff branches (foo/bar, X.Y.Z) etc and this does not seem to work for all the projects, then there should be some logic to handle the branch name that is passed as build param from the ci-man repo. 

I am not fully convinced that we need to trigger this job on the 'master' branch, and should be run when a stable/branch_name is created. I might be wrong. :-))

Patch-set: 1