Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> but why change things that were working?  they might not be optimal, sure, but is it worth the risk?

Actually the only reason I waded into this code in the first place was because of the tar -v. All this code to deal with the large numbers of files and then we echo the files to the log? The next thing I saw was the 'find -exec', it works but it does not scale and it is certainly less readable than 'mv $letter?* $letter'.

If someone were to request it, I would put this in another review. The case I make for keeping it in one change is the testing required did not really change because of these changes.

Patch-set: 1
Label: Code-Review=0