Trigger latest each time.
Trigger "latest" and "branch"

By default rtd serves documentation at /latest/

When a buid was triggered via curl, using a the old method.
"generic api integration"

curl -X POST -d "branches=${GERRIT_BRANCH}" -d "token=$RTD_TOKEN"
"$RTD_BUILD_URL

this returned for example:

{"build_triggered":true,"project":"o-ran-sc-doc","versions":["latest","master"]}

So the previous behavior was to build latest and master from a call with
branches=master

when we post via the v3 api.
lftools rtd project-build-trigger "$rtdproject" "$STREAM"

It actually respects the $STREAM variable and only builds "master"
This is not what we want,
we want latest to be updated as well.
so we need to trigger both builds.

There appears to be no way to disble 'latest'
and use only 'master'
and since this is the default behavior is seems best to
not fight against it, as to simplify initial project setup.

The only solution I see is to just always trigger latest
as well as the branch we want to build.

Signed-off-by: Aric Gardner <agardner@linuxfoundation.org>
Change-Id: I6e2062aa30383ee79a1d64d8f5327ef9a05cece6