Allow projects to change their default-version
Projects should be able to change the location
of the landing page. eg: from /latest/ to /branch-name/

set default version as latest, and allow projects
to change it via their jjb config.

Read the docs is built in a way that it
expects us to use a webhook that notifies read the docs of all
incomming changes, in this way it detects new tags.

If read the docs sees a tag, it creates a /stable/
directory, which follows the latest tag

I created an issue with them to clarify
whats going on

https://github.com/readthedocs/readthedocs.org/issues/6415

ISSUE: RELENG-2549
Signed-off-by: Aric Gardner <agardner@linuxfoundation.org>
Change-Id: I33a8df14e68340ba268c953a471daa02ee6fe766