Update patch set 8
Patch Set 8:

Hmm.

so, looks like the default behavior is this:
read the docs wants a /latest/ and /stable/

Stable follows /branch/ unless it sees a tag
then it follows the tag.

I dont really understand how the tag detection works
if you are not using a webhook
I have created and issue about this
https://github.com/readthedocs/readthedocs.org/issues/6415

So yeah, I'm glad you had a branch for me to poke at.
and a tag for that matter. sigh.

Patch-set: 8
Label: Code-Review=0