Update patch set 4
Patch Set 4:

You are asking for me to change underscores to hyphens in the name of a file that is used solely as an internal communication mechanism between two scripts launched by Jenkins.  I'm a little surprised that you care about punctuation so deeply! This change is a little out of scope for what I am trying to do here, but it would be trivial.

Patch-set: 4