Update patch set 13
Patch Set 13:

@chris so we can add the functionality to tag a ref on merge of the package-cloud release file.


But I have some questions.
This job produced the artifacts.
https://jenkins.o-ran-sc.org/job/com-log-docker-ci-pc-merge-master/12/console
But it was triggered by merging of the release file.
https://gerrit.o-ran-sc.org/r/c/com/log/+/1868

Seems oddly self referential. How did that work?

In this case which ref should have be tagged and signed?

Also, are you tagging on master, or only refs that exist on a release branch?

Patch-set: 13