Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Can we poll every couple of seconds or something to check the API's response code and find out when it's ready? Does it have a unique code or message for when it's not ready vs when it is complete?

No error code is returned from the lftools. yeah it would be good to poll Nexus on the status, if that is possible, since the time for each repo can vary depending on the size of the staging repo and the validation rules. 

I'd think this should be implemented in lftools and making it more robust, which would avoid shell scripting hacks.

Patch-set: 1