Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Can we poll every couple of seconds or something to check the API's response code and find out when it's ready? Does it have a unique code or message for when it's not ready vs when it is complete?
> 
> We can't trace to java threads to see when they are completed.
> We had the same problem here: https://gerrit.linuxfoundation.org/infra/c/releng/lftools/+/16448 
> 
> I don't understand why the maven jobs dont care about or cant see errors, and why we are uploading to nexus and asking nexus to do the job of validating the builds.

We basically need some kind of assurance that the stage repos are in a closed state (before proceeding with the release) and if not fail the maven-stage job that does the upload. On the day of the release we dont want to see surprises that the stage repo is not closed due to whatever issues and that could been identified during the maven-stage phase, and potentially delaying the release.

Patch-set: 1