Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> (1 comment)

Matching style is useful, but I would put it behind 'incremental improvement'. I could even make the case for:
git_tag=$GIT_TAG
unset GIT_TAG
if the GIT_TAG is not required by any command the script calls. Which is the case for GIT_TAG. It is passed as an argument, but it is NOT need to be passed in the environment to any child process. When you are writing 20 lines scripts for yourself this is NOT important, but if you are writing large complex production scripts, we need to worry about this kind of stuff. I won't stop you merge this as-is, but I just think we can do better.

If GIT_TAG is supposed to be set by the YAML then your script should error out. That validates that the JJB/YAML is correct. If the scripts have defaults for GIT_TAG the we cannot validate the YAML. Classic example of the 'fault tolerance' and that makes our environment hard to maintain. ERROR: GIT_TAG is easy to fix, a python stack trace due because some fault when a bad value of GIT_TAG is referenced, not so much.

Patch-set: 3
Reviewer: Gerrit User 182 <182@f2a2391d-0403-4fa5-b183-4056999a8a09>