Update patch set 4
Patch Set 4:

@Aric, to answer your question, why split git tag from version?

ORAN has a repository that releases a PyPI distribution and PackageCloud (shared library as DEB/RPM) artifacts. The version strings on those should match, because they work hand-in-glove with each other -- the Python is a binding for the C library. But the project wishes to tag the two project releases differently.  Adding the optional git_tag entry in the release yaml file lets them tag the git repo with "deb-2.0.0" in the deb/rpm release job, and "py-2.0.0" in the PyPI release job. 

Does this explain the need?

Patch-set: 4
Reviewer: Gerrit User 114 <114@f2a2391d-0403-4fa5-b183-4056999a8a09>