Update patch set 20
Patch Set 20:

> Patch Set 19: Code-Review-1
> 
> (1 comment)

The 'cost' command is included in the v0.29.0 release of lftools. Of the scripts that install/use lftools, none of them are tied to a specific release. So they they most use lftools 'master', which is why this issue did not effect any of my testing. There is always a chance that some script in some projects does install a specific version of lftools that is less than v0.29.0. I will add a note to the release note.

Patch-set: 20
Reviewer: Gerrit User 182 <182@f2a2391d-0403-4fa5-b183-4056999a8a09>