Example of using generic-webhook-trigger
Current github merge job runs on a cron
and can't be configured to only run when
a matched file is merged.

triggers:
  - github
  - pollscm:
      cron:

generic-webhook-trigger allows this,
but does not allow remerge with a trigger phrase
trigger-phrase (string)

This has been tested in releated INFO file work,
and I will use this patchset to test against
release files

I need
generic-webhook-trigger installed on ODL
for this to verify.

ISSUE-ID: RELENG-1864
Signed-off-by: Aric Gardner <agardner@linuxfoundation.org>
Change-Id: Ib156c10a8a6bdc4a9bd3652e1ce6febc99372e8f