Update patch set 1
Patch Set 1:

Tested newly tweaked javadoc-verify jobs both with and without mvn-subdir on acumos/common-dataservice in the Jenkins sandbox.  Only the generate task was affected, which is part of both verify and merge jobs. The publish task (merge only) was not affected. Altho I think those jobs and their logs will be  gone long before anyone reviews this, here are the links:

With non-default subdir:

https://jenkins.acumos.org/sandbox/job/common-dataservice-maven-javadoc-verify-master-openjdk11/1/

With default subdir ".":

https://jenkins.acumos.org/sandbox/job/common-dataservice-maven-javadoc-verify-master-openjdk11/2/

Patch-set: 1