Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> Do you plan to use lf-venv to replace the function in global-jjb/shell/common-variables.sh anytime soon? Or maybe revise global-jjb/shell/maven-deploy.sh to call lf-venv directly instead of this function?
> 
>   # Activates the lftools virtualenv
>   lftools_activate() {
>     virtualenv --quiet "/tmp/v/lftools"
>     set +u  # Ignore unbound variables in activate
>     # shellcheck source=/tmp/v/lftools/bin/activate disable=SC1091
>     source "/tmp/v/lftools/bin/activate"
>     set -u  # Restore unbound variable checking
>   }

The plan is to delete common-variables.sh. It is one of the files shared by many jobs by using YAML/JJB to 'concatenate' this script with other scripts into a single Jenkins script  (in lieu of calling a function that is defined in lf-env.sh). I have already remove a few references to this file, more will happen over time. I will definitely fight anyone attempting to add new references to this file.

Patch-set: 6