Update patch set 1
Patch Set 1:

Anil:
There are a number of comments about a 'work-around' for a bug in 'git review (v1.24)'. Currently ODL is using git-review v1.27 so I think we could remove the 'work-around' as part of this cleanup. I don't understand what changes are needed to remove the work-around. Git blames you for this code, so I assume you have some idea what needs to be changed.

Patch-set: 1