Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> (1 comment)

When you really get down to it, when the command supports a _OPTIONS env var we depend on the command to correctly deal with any conflicts. Only the command can do that (because it knows what all the options/flags are and how they interact). We can only hope the command does it's job (and returns non-zero status when it finds a problem).

PS This code needs to be rebased

Patch-set: 5
Reviewer: Gerrit User 182 <182@f2a2391d-0403-4fa5-b183-4056999a8a09>