Update patch set 7
Patch Set 7: Code-Review+1

This is sooo messed up. I did some digging into maven and the way we are using MAVEN_OPTS, MAVEN_OPTIONS & MAVEN_PARAM.  MAVEN_OPTS is an environmental variable that recognized by maven. From the maven documentation:
    MAVEN_OPTS: Specify additional options using the MAVEN_OPTS environment variable. It is for
    passing parameters to the Java VM when running Maven.

MAVEN_PARAM is a env shell variables set in JJB config but it has no effect on maven unless it is explicitly added to the command line.

MAVEN_OPTIONS is a shell variable that is defined in common variables.sh. Sometimes it is exported (why??) but it has no effect on maven unless it is explicitly added to the command line.

MAVEN_OPTIONS & common-variables.sh are already on my radar and I have plans to deprecate/delete the file (it is one of the those scripts that is concatenated with other scripts).

Patch-set: 7
Reviewer: Gerrit User 182 <182@f2a2391d-0403-4fa5-b183-4056999a8a09>
Label: Code-Review=+1