Update patch set 9
Patch Set 9:

Forgive me for resurrecting this very old change, but I'd like to understand it a little better. 

First, java-opts is wholly undocumented in the RST, why was this merged without a description?

Second, why is this environment varaiable set and passed along only on the FIRST maven invocation, the one that invokes a goal like "clean install" and not on the SECOND maven invocation, the one that calls sonar plugin?

Third, the JVM doc says that _JAVA_OPTIONS is somewhat deprecated:  https://stackoverflow.com/questions/17781405/information-about-java-options     Also Tim J to his credit flagged the same thing at the time.

The Jira issue doesn't shed a lot of light, only that Suresh made something work, somehow, with _JAVA_OPTIONS.

Thanks in advance.

Patch-set: 9
CC: Gerrit User 114 <114@f2a2391d-0403-4fa5-b183-4056999a8a09>