Update patch set 2
Patch Set 2:

Now that I check the usage, I see that the maven-javadoc-generate.sh script is the second in a block of catenated shell-script lines in Jenkins.  The -l option on the first block, shell script common-variables.sh, is probably honored and that sets login-shell status.

So unless you all think it's helpful to make a change to -l, I'm just gonna leave the -l here for now.

Maybe when Tim gets around to expunging common-variables.sh this should be revisited.

Patch-set: 2