Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Now that I check the usage, I see that the maven-javadoc-generate.sh script is the second in a block of catenated shell-script lines in Jenkins.  The -l option on the first block, shell script common-variables.sh, is probably honored and that sets login-shell status.
> 
> So unless you all think it's helpful to make a change to -l, I'm just gonna leave the -l here for now.
> 
> Maybe when Tim gets around to expunging common-variables.sh this should be revisited.

common-variables.sh (and all the other 'concatenated' scripts) are on my list all right. The '#!/bin/bash -l' is just a comment, unless it is on the first line of the script. so when you concatenate the scripts, only the first one counts. I say delete, if it had not been there, we never would have had this discussion, plus it will make it easier when I go to unravel this mess.

Patch-set: 2
Reviewer: Gerrit User 182 <182@f2a2391d-0403-4fa5-b183-4056999a8a09>