Update patch set 3
Patch Set 3:

What this warning is saying is that not having the venv bin in the path can cause problems. It is a warning because it is OK to install but it needs to be in the PATH at runtime. I am pretty sure you only get this warning with user venv's.

In a 'user' venv' the libraries are always in the python search path even if it is not in your PATH. What this means if you execute x from anywhere it will still search for python libraries in ~/.local before it searches for the libraries in system. So if you get a command from outside the user venv it will still use libraries from the user venv. That is why it warns about PATH.

If we were installing all python packages in the same venv, a user venv might be an option. Unfortunately, there are a number of reason's why we don't want to go that route.

A non 'user' venv does NOT work the same way. If you run a command from a venv it will use libraries that are under control of the venv.

Patch-set: 3
CC: Gerrit User 182 <182@f2a2391d-0403-4fa5-b183-4056999a8a09>