Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+1
> 
> LGTM, please add a missing release note file

I don't think a release note is necessary here. No functionality is changed and no user facing features are changed either. I'm not sure why an end user reading the release notes would care about a CI change that only affects the maintainers of this project?

It's purely moving shellcheck from coala to pre-commit all 3 of which are tools we are already using in this repo.

Patch-set: 2