Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> I prefer to have it as it makes debugging jjb a lot easier when there's an issue. Without the debug flag we're often left guessing which file JJB was last parsing before it threw an error. This is one of those cases where it seems like a lot of noise until something fails and someone has to trace back the error then you wish you had it.
> 
> With that said I'm not really spending much time these days debugging JJB files for the community so if folks think having less logs is useful feel free to ignore my vote.

It makes little sense to make 99.9% case more difficult to make it easier for the .1% case. Particularly when it is so easy to edit the tox.ini to enable it when it is needed. The output of tox run goes from 1900 -> 500 lines. The shellcheck errors are right at the bottom, but issues from other checks get intermixed in the output.

Patch-set: 2