Update patch set 2
Patch Set 2:

> Patch Set 2:
> It makes little sense to make 99.9% case more difficult to make it easier for the .1% case. 

I have to disagree here. 99% of the time, I don't actually care about the logs. If Jenkins gives a verified +1 vote I never read the logs in that case.

The only time I ever read logs at all is when there is an error. So in that case I want to have DEBUG logs available to me so that I can appropriately troubleshoot the problem.

> Particularly when it is so easy to edit the tox.ini to enable it when it is needed. The output of tox run goes from 1900 -> 500 lines. The shellcheck errors are right at the bottom, but issues from other checks get intermixed in the output.

This sounds like a different problem then JJB printing too much info. It's a problem of how to parse out the different tox logs especially since we run tox in parallel mode. We do archive tox logs separately on the log server so if you look at for example:

https://logs.opendaylight.org/releng/vex-yul-odl-jenkins-1/lf-infra-global-jjb-tox-verify-any/2451/tox/

You'll see each tox environment has a separate folder generating logs related to that one tox virtualenv.

In reviewing this now though it looks like the logs do not contain the stdout and stderr of the commands being run, it used to. Something must have changed with our tox scripts that removed the stdout and stderr logs for each environment. We should investigate turning it back on so that you can get the log separation you're looking for.

Patch-set: 2
Reviewer: Gerrit User 8 <8@f2a2391d-0403-4fa5-b183-4056999a8a09>