Update patch set 2
Patch Set 2:

Another route we could go is if we upgrade all of our base python's to python3 and use the latest version of tox. We can use:

    tox -p auto

In thise mode tox will run all of the virtualenvs in parallel BUT only prints the log IF the commands fail. Otherwise every environment only gets a single line output to state that the command ran successfully. Here's an example:

 $ tox -p auto
 ✔ OK prefix in 0.707 second
 ✔ OK jjb-version in 0.717 second
 ✖ FAIL robot in 0.752 second
 Failed robot under process 74782, stdout:
 robot run-test-pre: PYTHONHASHSEED='2515262986'
 robot runtests: commands[0] | python 
 /Users/zxiiro/src/git.opendaylight.org/releng/builder/check_robot.py
 ERROR: InvocationError for command 
 '/Users/zxiiro/src/git.opendaylight.org/releng/builder/.tox/robot/bin/python 
 /Users/zxiiro/src/git.opendaylight.org/releng/builder/check_robot.py' (exited with code 1)
 ✔ OK docs in 3.591 seconds
 ✔ OK pre-commit in 4.018 seconds
 ✔ OK coala in 16.757 seconds
 _______________________________________________________ summary 
 ________________________________________________________
   coala: commands succeeded
   docs: commands succeeded
   jjb-version: commands succeeded
   pre-commit: commands succeeded
   prefix: commands succeeded
 ERROR:   robot: parallel child exit code 1

I think this is an improvement that would be worth pursuing if you want to cut down on the spam of log outputs since in this case we can keep the DEBUG level log output but it will only be pasted to the console _if_ a failure is detected rather than all the time.

Patch-set: 2
Reviewer: Gerrit User 8 <8@f2a2391d-0403-4fa5-b183-4056999a8a09>